% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.fixed.point}
\alias{plot.fixed.point}
\title{Plot a \code{FEM.time} object at a fixed spatial point in a given time interval}
\usage{
plot.fixed.point(x,x1,x2,x3=NULL,timeInterval = x$mesh_time, N=100, NameLocation=NA,type="l", col='red', samePlot = FALSE, main = NA, ...)
}
\arguments{
\item{x}{A \code{FEM.time} object.}

\item{x1}{First coordinate of the point.}

\item{x2}{Second coordinate of the point.}

\item{x3}{Third coordinate of the point.}

\item{timeInterval}{vector of time points where the estimated field has to be evaluated}

\item{N}{Number of time istants in range for the plot}

\item{NameLocation}{Name of selected location (for main parameter of plot)}

\item{...}{Arguments representing graphical options}
}
\description{
Plot of a \code{FEM.time} object, generated by \code{FEM.time} or returned by \code{smooth.FEM.time}
}
\seealso{
\code{\link{FEM.time}}
}
