% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mesh.3D}
\alias{plot.mesh.3D}
\title{Plot a mesh.3D object}
\usage{
\method{plot}{mesh.3D}(mesh, ...)
}
\arguments{
\item{mesh}{A \code{mesh.3D} object generated by \code{create.mesh.3D}.}

\item{...}{Arguments representing graphical options to be passed to \link[graphics]{par}.}
}
\description{
Plot a \code{mesh.3D} object, generated by \code{create.mesh.3D}.
}
\examples{
library(fdaPDE)

##Load a mesh.3D object
data(sphere3D)
##Plot the triangulation of the object
plot.mesh.3D(sphere3D)
}
