% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.R
\name{points.projection.2.5D}
\alias{points.projection.2.5D}
\title{Project 3D points onto 2D 2.5D triangular mesh}
\usage{
\method{points}{projection.2.5D}(mesh, locations)
}
\arguments{
\item{mesh}{A mesh.2.5D object representing the triangular mesh, created by \link{create.mesh.2.5D}.}

\item{locations}{3D points to be projected onto 2.5D triangular mesh.}
}
\value{
3D points projected onto 2.5D triangluar mesh.
}
\description{
This function projects any 3D points onto 2.5D triangular mesh.
}
\examples{
library(fdaPDE)

## Upload the hub2.5D the data
data(hub25Ddata)

## Create mesh
mesh <- create.mesh.2.5D(nodes = nodes,triangles = triangles)

# Create 3D uniform grid points to be projected
x <- seq(-3,3,by=0.6)
y <- seq(-3,3,by=0.6)
 z <- seq(-3,3,by=0.6)
grid = expand.grid(x=x, y=y, z=z)
loc = points.projection.2.5D(mesh, grid)
}
